Printer Issue

+++ src/utils.mjs
@@ -265,7 +265,7 @@ ${normalizedFoodList.map(({ name, modifier, num, price }) => `|${name} |\n${modi
   const remarkMd = remark ? `Remark: |${remark}\n` : ''
   const trainingLevy = `Training Levy 2%: | ${0.02 * parseFloat(totalPrice.replace(/,/g, ""))}`
   const serviceCharge = `Service Charge 5%: | ${0.05 * parseFloat(totalPrice.replace(/,/g, ""))}`
-  const itemValue = `Item Value: |${parseFloat(totalPrice.replace(/,/g, "")) - (0.16 * parseFloat(totalPrice.replace(/,/g, ""))) - (0.05 * parseFloat(totalPrice.replace(/,/g, ""))) - (0.02 * parseFloat(totalPrice.replace(/,/g, "")))}`
+  const itemValue = `Item Value: |${parseInt(totalPrice.replace(/,/g, "")) - (0.16 * parseFloat(totalPrice.replace(/,/g, ""))) - (0.05 * parseFloat(totalPrice.replace(/,/g, ""))) - (0.02 * parseFloat(totalPrice.replace(/,/g, "")))}`
   const vatValue = `VAT 16%: | ${0.16 * parseFloat(totalPrice.replace(/,/g, ""))}`
